+++ src/components/shared/index.ts
@@ -1,2 +1,3 @@
 export * from "./onboarding-navbar/onboarding-navbar";
 export * from "./form/form";
+export * from "./sign-up-step";
