+++ src/components/shared/sign-up-step/index.ts
@@ -1,0 +1,1 @@
+export * from "./type-birth-date-step";
